Change type of pvp_score to i64 #LABTROYALC-3 @author Sergej

+++ clicker_backend/src/main.rs
@@ -324,7 +324,7 @@ async fn load_game_state_from_database(id: i64, pool: &PgPool) -> GameState {
     }
 }
 
-async fn search_for_enemy(pvp_score: i32, pool: &PgPool) -> i64 {
+async fn search_for_enemy(pvp_score: i64, pool: &PgPool) -> i64 {
     println!("Searching for Enemy");
     match sqlx::query!(
         "SELECT id, game_state FROM Player WHERE pvp_score <= $1 ORDER BY pvp_score ASC",
